Rename Y::ActionCable::UpdateLog to Y::UpdateLog
The update log is transport-agnostic ActiveRecord storage; nothing about
it is ActionCable. The module moves to lib/y/update_log.rb as Y::UpdateLog,
still shipping in the yrby-actioncable gem (added to its file list,
excluded from core's), still required by y/action_cable. The changelog
entry now describes the module and the generator separately.
