+++ CHANGELOG-actioncable.md
@@ -8,15 +8,21 @@ this project aims to follow [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `rails generate yrby:install`: wires a Rails app in one step — a
-  `DocumentChannel` speaking the y-websocket protocol, an
-  ActiveRecord-backed durable store, and its migration. The store is an
-  append-only update log with inline compaction (`compact_every`, default
-  500): rows collapse into one snapshot row so `on_load` replays the
-  compaction window, not the document's full history. The generated files
-  are plain app code, and the store's merge goes through
-  `compacted_state_update` so a gappy recorded update can never be served
-  to peers. Store behavior is pinned by tests against a real database.
+- `Y::UpdateLog`: durable storage for collaborative documents as a module
+  any ActiveRecord model with `payload`/`document_key` columns includes.
+  An append-only update log with inline compaction (`compact_every`,
+  default 500): rows collapse into one snapshot row so `on_load` replays
+  the compaction window, not the document's full history. Merges go
+  through `compacted_state_update` so a gappy recorded update can never
+  be served to peers, and compaction skips a document holding a pending
+  (causally-gapped) update rather than deleting the only healable copy.
+  Behavior is pinned by tests against a real database.
+- `rails generate yrby:install [ModelName]`: wires a Rails app in one
+  step — a `DocumentChannel` speaking the y-websocket protocol, an
+  update-log model (`include Y::UpdateLog`), and its migration. The
+  optional argument names the model, and the table and migration derive
+  from it; namespaced names are rejected (the inferred table would miss
+  the generated migration's table).
 
 ## [0.3.1] - 2026-07-01
