+++ lib/generators/yrby/install/USAGE
@@ -1,7 +1,7 @@
 Description:
     Wires the app for collaborative documents: a DocumentChannel speaking
     the y-websocket protocol, an update-log model backed by
-    Y::ActionCable::UpdateLog (append-only rows with compaction), and its
+    Y::UpdateLog (append-only rows with compaction), and its
     migration.
 
 The optional argument names the update-log model; the table and the
