+++ lib/generators/yrby/install/templates/yrby_document_update.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 # One CRDT delta (or compacted snapshot) per row for a collaborative
-# document. Y::ActionCable::UpdateLog provides load/append/compact! and
+# document. Y::UpdateLog provides load/append/compact! and
 # inline compaction (tune with `self.compact_every = ...`); rows only
 # accumulate between compactions.
 class <%= model_class_name %> < ApplicationRecord
-  include Y::ActionCable::UpdateLog
+  include Y::UpdateLog
 end
