+++ lib/y/action_cable.rb
@@ -16,4 +16,4 @@ module Y
 end
 
 require "y/action_cable/sync"
-require "y/action_cable/update_log"
+require "y/update_log"
